Auto-generate tool list
At present, adding a new tool to Quality would require updating
several places in the source code. By namespacing the tools module
and with a little metaprogramming we can reduce by one step.
